refactor: package 경로 재설정
