+++ src/main/java/cart/catalog/controller/ProductListController.java
@@ -1,25 +1,25 @@
-package cart.product.controller;
+package cart.catalog.controller;
 
-import cart.product.service.ProductListService;
+import cart.catalog.service.ProductListService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
 public class ProductListController {
-    
+
     private final ProductListService productListService;
-    
+
     public ProductListController(final ProductListService productListService) {
         this.productListService = productListService;
     }
-    
+
     @GetMapping("/")
     public String renderProductListPage(final Model model) {
         model.addAttribute("products", this.productListService.display());
         return "index";
     }
-    
+
     @GetMapping("/admin")
     public String renderAdminPage(final Model model) {
         model.addAttribute("products", this.productListService.display());
